removed class from namespace with same name fixed namespace for all scripts moved NameGenerator to Modules rename namspace "player" to plural to prevent name conflicts with class
