--- Assets/Scripts/Base/Assets/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b4a2010173b4aa34797056e79f1a694c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
