+++ Assets/Scripts/Base/GameManager.cs
@@ -1,47 +1,50 @@
 ﻿using System;
 using UnityEngine;
-using Game.State;
+using Base.State;
 
-public class GameManager : MonoBehaviour
+namespace Base
 {
-    private IGameState _gameState;
-    private GameObject _mainMenuUI;
-    
-    private void Start()
+    public class GameManager : MonoBehaviour
     {
-        _mainMenuUI = GameObject.Find("MainMenu");
-    }
+        private IGameState _gameState;
+        private GameObject _mainMenuUI;
 
-    // Update is called once per frame
-    void Update()
-    {
-        Console.WriteLine("[GameManager] inside");
-        if (_gameState != null)
+        private void Start()
         {
-            _gameState = _gameState?.update();
+            _mainMenuUI = GameObject.Find("MainMenu");
         }
-        else if(_gameState == null && !_mainMenuUI.activeSelf)
+
+        // Update is called once per frame
+        void Update()
         {
-            _mainMenuUI.SetActive(true);
-            Debug.Log("[GameManager] reenabled Menu");
+            Console.WriteLine("[GameManager] inside");
+            if (_gameState != null)
+            {
+                _gameState = _gameState?.update();
+            }
+            else if (_gameState == null && !_mainMenuUI.activeSelf)
+            {
+                _mainMenuUI.SetActive(true);
+                Debug.Log("[GameManager] reenabled Menu");
+            }
         }
-    }
 
-    public void SetupGame(RuleBook ruleBook)
-    {
-        if (_gameState == null)
+        public void SetupGame(RuleBook ruleBook)
         {
-            Console.WriteLine("[GameManager] create new Game of type " + ruleBook.ruleBookTitle);
-            
-            SetupState setupState = new SetupState();    //todo pass RuleBook to setup which is chosen in the menu by the player
-            PlayState playState = new PlayState();
-            EndState endState = new EndState();
+            if (_gameState == null)
+            {
+                Console.WriteLine("[GameManager] create new Game of type " + ruleBook.ruleBookTitle);
 
-            setupState.SetPlayState(playState);
-            playState.SetEndState(endState);
-            endState.setSetupState(setupState);
+                SetupState setupState = new SetupState();    //todo pass RuleBook to setup which is chosen in the menu by the player
+                PlayState playState = new PlayState();
+                EndState endState = new EndState();
 
-            _gameState = setupState;
+                setupState.SetPlayState(playState);
+                playState.SetEndState(endState);
+                endState.setSetupState(setupState);
+
+                _gameState = setupState;
+            }
         }
     }
 }
