+++ Assets/Scripts/Modules/DicingZombies/Assets/Dice/EasyZombieDice.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
-using Play.DicingZombies.Assets.Dice;
 
-namespace Play.DicingZombies
+namespace Modules.DicingZombies.Assets.Dice
 {
     public class EasyZombieDice : ZombieDice
     {
