+++ Assets/Scripts/Modules/DicingZombies/Assets/Dice/ZombieDiceSideValue.cs
@@ -1,6 +1,4 @@
-using System;
-
-namespace Play.DicingZombies
+namespace Modules.DicingZombies.Assets.Dice
 {
     public class ZombieDiceSideValue : DiceSideValue
     {
